+++ physics_engine/physics_engine/AABB.h
@@ -18,15 +18,15 @@ public:
      *  @brief  Returns the dimensions of the box in terms of half-lengths.
      *  @return The dimensions of the box along the three axes in half-lengths. I.e. distance from the centre to the end.
      */
-    glm::vec3 getDim();
+    glm::dvec3 getDim();
 
     /*!
      *  @brief  Sets the dimensions of the box.
      *  @param[in]  dim The dimensions to set the box to.
      */
-    void setDim(glm::vec3& dim);
+    void setDim(glm::dvec3& dim);
 
 private:
-    glm::vec3   mDim;   /// Dimensions of the box along the three axes. Half lengths.
+    glm::dvec3   mDim;   /// Dimensions of the box along the three axes. Half lengths.
 };
